refactoring trusted orgs section
